added additional/grouped GraphQLComponent API tests, cleaned up resolver import and removed associated uneeded abstract resolver test, move delegateToComponent and associated tests to own folder/module and updated references in resolvers and index.js
