+++ lib/__tests__.js
@@ -8,317 +8,134 @@ const GraphQLComponent = require('./index');
 
 Test('component API', (t) => {
 
-  t.plan(1);
+  t.test('component name (anonymous constructor)', (st) => {
+    const component = new GraphQLComponent();
+    t.equals(component.name, 'GraphQLComponent', `unnamed constructor results in component named 'GraphQLComponent'`);
+    st.end();
+  })
 
-  const component = new GraphQLComponent();
-
-  t.ok(component._id, '_id populated.');
-});
-
-Test('component isComponent', (t) => {
-
-  t.test('isComponent not a subclass', (t) => {
-    t.plan(1);
-
-    t.ok(!GraphQLComponent.isComponent(Object.create({ types: [], resolvers: {} })), 'not a subclass');
+  t.test('component name (named constructor)', (st) => {
+    class Named extends GraphQLComponent {}
+    const component = new Named();
+    t.equals(component.name, 'Named', `named constructor results in 'Named'`);
+    st.end();
   });
 
-  t.test('isComponent', (t) => {
-    t.plan(1);
-
-    t.ok(GraphQLComponent.isComponent(new GraphQLComponent()), 'new super class is component');
+  t.test('component id', (st) => {
+    const component = new GraphQLComponent();
+    t.ok(component.id, `got component's id`);
+    st.end();
   });
 
-  t.test('isComponent subclass', (t) => {
-    t.plan(1);
-
-    t.ok(GraphQLComponent.isComponent(new class extends GraphQLComponent { }), 'new subclass is component');
+  t.test('isComponent with config object', (st) => {
+    st.notOk(GraphQLComponent.isComponent(Object.create({ component: new GraphQLComponent, exclude: ['Query.a'] })), 'is not a component');
+    st.end();
   });
 
-});
+  t.test('isComponent with new base class instance', (st) => {
+    console.log(typeof new GraphQLComponent());
+    st.ok(GraphQLComponent.isComponent(new GraphQLComponent()), 'is a component');
+    st.end();
+  });
 
-Test('component resolver delegate', async (t) => {
+  t.test('isComponent with new subclass', (st) => {
+    t.ok(GraphQLComponent.isComponent(new class extends GraphQLComponent { }), 'is a component');
+    st.end();
+  });
 
-  t.plan(1);
+  t.test('component context', (st) => {
+    const component = new GraphQLComponent();
+    const context = component.context;
+    st.ok(typeof context === 'function', 'context is a function');
+    st.ok(typeof context.use === 'function', 'context has a use funtion');
+    st.end();
+  });
 
-  const component = new GraphQLComponent({
-    imports: [
-      new GraphQLComponent({
-        types: `
-          type Test {
-            value: Boolean
-          }
-          type Query {
-            test: Test
-          }
-        `,
+  t.test('component types', (st) => {
+    const component = new GraphQLComponent({
+      types: `type Query { a: String }`,
+      imports: [new GraphQLComponent({
+        types: `type Query { b: B} type B { someField: String}`}
+      )]
+    });
+
+    st.deepEquals(component.types, [`type Query { a: String }`], `only the component's own types are returned (no imports)`);
+    st.end();
+  });
+
+  t.test('component resolvers', (st) => {
+    const component = new GraphQLComponent({
+      resolvers: {
+        Query: {
+          a() { return 'hello'}
+        }
+      },
+      imports: [new GraphQLComponent({
         resolvers: {
           Query: {
-            test() {
-              return {
-                value: true
-              }
+            b() {
+              return 'goodbye';
             }
           }
         }
-      })
-    ]
+      })]
+    });
+
+    st.equals(Object.keys(component.resolvers.Query).length, 1, `only the component's own resolvers are returned`);
+    st.end();
   });
 
-  const { data } = await graphql.execute({
-    document: gql`query { test { value } }`,
-    schema: component.schema,
-    rootValue: undefined,
-    contextValue: {}
-  });
-  
-  t.equal(data.test.value, true, 'resolved');
-});
-
-Test('component resolver delegate errors', async (t) => {
-
-  t.plan(3);
-
-  const component = new GraphQLComponent({
-    imports: [
-      new GraphQLComponent({
-        types: `
-          type Test {
-            value: Boolean
-            err: Boolean
-          }
-          type Query {
-            test: Test
-          }
-        `,
-        resolvers: {
-          Query: {
-            test() {
-              return {
-                value: true,
-                err: true
-              }
-            }
-          },
-          Test: {
-            err(_) {
-              if (_.err) {
-                throw new Error('error');
-              }
-            }
-          }
-        }
-      })
-    ]
+  t.test('component imports', (st) => {
+    const childThatAlsoHasImports = new GraphQLComponent({
+      types: `type Query { c: String }`,
+      resolvers: { Query: { c() { return 'hello' }}},
+      imports: [new GraphQLComponent()]
+    })
+    const root = new GraphQLComponent({
+      imports: [
+        childThatAlsoHasImports
+      ]
+    });
+    st.equals(root.imports.length, 1, `only component's own imports are returned`);
+    st.equals(childThatAlsoHasImports.id, root.imports[0].component.id, `id of only import matches import instance's id`);
+    st.end();
   });
 
-  const document = gql`
-    query { 
-      foo: test { value, err } 
-      bar: test { value }
-    }`;
+  t.test('component mocks', (st) => {
+    const component = new GraphQLComponent({
+      mocks: { Query: { mockA() { return 'mockAString' }}},
+      imports: [new GraphQLComponent({
+        mocks: { Query: { mockB() { return 'mockBString' }}}
+      })]
+    });
 
-  const { data, errors } = await graphql.execute({
-    document,
-    schema: component.schema,
-    rootValue: undefined,
-    contextValue: {}
+    st.equals(Object.keys(component.mocks).length, 1, `only component's own mocks are returned`);
+    st.end();
   });
 
-  t.equal(data.foo.value, true, 'resolved alias 1');
-  t.equal(data.bar.value, true, 'resolved alias 2');
-  t.ok(errors && errors.length > 0, 'got error');
-});
+  t.test('component directives', (st) => {
+    const component = new GraphQLComponent({
+      directives: { parentDirective: () => {}},
+      imports: [new GraphQLComponent({
+        directives: { childDirective: () => {}}
+      })]
+    });
 
-Test('delegateToComponent from root type', async (t) => {
-
-  t.plan(3);
-
-  const childComponent = new GraphQLComponent({
-    types: `
-      type Child {
-        childField: String
-        anotherChildField: String
-      }
-      type Query {
-        childById: Child
-      }
-    `,
-    resolvers: {
-      Query: {
-        childById() {
-          return {
-            childField: 'Child Field',
-            anotherChildField: 'Another Child Field'
-          }
-        }
-      }
-    }
+    st.equals(Object.keys(component.directives).length, 1, `only component's own directives are returned`);
+    st.end();
   });
 
-  const component = new GraphQLComponent({
-    types: `
-      type Parent {
-        child: Child
-      }
-      type Child {
-        addedField: String
-      }
-      type Query {
-        parent: Parent
-      }
-    `,
-    resolvers: {
-      Query: {
-        parent: async function (_, args, context, info) {
-          const child = await GraphQLComponent.delegateToComponent(childComponent, {
-            fieldMap: {
-              child: 'childById'
-            },
-            subPath: 'child',
-            contextValue: context,
-            info
-          });
+  t.test('component datasources', (st) => {
+    const component = new GraphQLComponent({
+      dataSources: ['parentDataSourcePlaceHolder'],
+      imports: [new GraphQLComponent({
+        dataSources: ['childDataSourcePlaceHolder']
+      })]
+    });
 
-          return {
-            child
-          };
-        }
-      },
-      Child: {
-        addedField() {
-          return 'Added from Parent'
-        }
-      }
-    },
-    imports: [
-      childComponent
-    ]
+    st.equals(Object.keys(component.dataSources).length, 1, `only component's own dataSources are returned`);
+    st.end();
   });
-
-  const document = gql`
-    query { 
-      parent1: parent {
-        child {
-          childField
-          addedField
-        }
-      }
-      parent2: parent {
-        child {
-          anotherChildField
-          addedField
-        }
-      }
-    }`;
-
-  const result = await graphql.execute({
-    document,
-    schema: component.schema,
-    rootValue: undefined,
-    contextValue: {}
-  });
-
-  t.ok(!result.errors, 'no errors');
-  
-  const { parent1, parent2 } = result.data;
-
-  t.deepEqual(parent1, { child: { childField: 'Child Field', addedField: 'Added from Parent' }}, 'received correct first result');
-  t.deepEqual(parent2, { child: { anotherChildField: 'Another Child Field', addedField: 'Added from Parent' }}, 'received correct second result');
-});
-
-Test('delegateToComponent from type resolver', async (t) => {
-
-  t.plan(3);
-
-  const childComponent = new GraphQLComponent({
-    types: `
-      type Child {
-        childField: String
-        anotherChildField: String
-      }
-      type Query {
-        child: Child
-      }
-    `,
-    resolvers: {
-      Query: {
-        child() {
-          return {
-            childField: 'Child Field',
-            anotherChildField: 'Another Child Field'
-          }
-        }
-      }
-    }
-  });
-
-  const component = new GraphQLComponent({
-    types: `
-      type Parent {
-        child: Child
-      }
-      type Child {
-        addedField: String
-      }
-      type Query {
-        parent: Parent
-      }
-    `,
-    resolvers: {
-      Query: {
-        parent: async function () {
-          return {};
-        }
-      },
-      Parent: {
-        child(_, args, context, info) {
-          return GraphQLComponent.delegateToComponent(childComponent, {
-            contextValue: context,
-            info
-          });
-        }
-      },
-      Child: {
-        addedField() {
-          return 'Added from Parent'
-        }
-      }
-    },
-    imports: [
-      childComponent
-    ]
-  });
-
-  const document = gql`
-    query { 
-      parent1: parent {
-        child {
-          childField
-          addedField
-        }
-      }
-      parent2: parent {
-        child {
-          anotherChildField
-          addedField
-        }
-      }
-    }`;
-
-  const result = await graphql.execute({
-    document,
-    schema: component.schema,
-    rootValue: undefined,
-    contextValue: {}
-  });
-
-  t.ok(!result.errors, 'no errors');
-  
-  const { parent1, parent2 } = result.data;
-
-  t.deepEqual(parent1, { child: { childField: 'Child Field', addedField: 'Added from Parent', }}, 'received correct first result');
-  t.deepEqual(parent2, { child: { anotherChildField: 'Another Child Field', addedField: 'Added from Parent' }}, 'received correct second result');
 });
 
 Test('mocks', async (t) => {
@@ -489,136 +306,3 @@ Test('integration: data source', (t) => {
   
 });
 
-Test('proxy to child with root query that returns interface type (using __resolveType())', async (t) => {
-  const child = new GraphQLComponent({
-    types: `
-      type Query {
-        things: [Thing]
-      }
-      interface Thing {
-        id: ID
-      }
-      type Person implements Thing {
-        id: ID
-        name: String
-      }
-      type Animal implements Thing {
-        id: ID
-        someField: Int
-      }
-    `,
-    resolvers: {
-      Query: {
-        things() {
-          return [
-            {
-              id: '1',
-              name: 'Joe Smith'
-            }
-          ]
-        }
-      },
-      Thing: {
-        __resolveType(parent) {
-          if (parent.name) {
-            return 'Person';
-          }
-          return 'Animal';
-        }
-      }
-    }
-  });
-
-  const parent = new GraphQLComponent({
-    imports: [
-      child
-    ]
-  });
-
-  const result = await graphql.execute({
-    document: gql`
-      query {
-        things {
-          id
-          ... on Person {
-            name
-          }
-        }
-      }
-    `,
-    schema: parent.schema,
-    contextValue: {}
-  });
-
-  t.deepEquals(result.data.things, [{ id: '1', name: 'Joe Smith', __typename: 'Person' }], 'interface type resolved')
-  t.notOk(result.errors, 'no errors');
-  t.end();
-});
-
-Test('proxy to child with root query that returns interface type (using __isTypeOf())', async (t) => {
-  const child = new GraphQLComponent({
-    types: `
-      type Query {
-        things: [Thing]
-      }
-      interface Thing {
-        id: ID
-      }
-      type Person implements Thing {
-        id: ID
-        name: String
-      }
-      type Animal implements Thing {
-        id: ID
-        someField: Int
-      }
-    `,
-    resolvers: {
-      Query: {
-        things() {
-          return [
-            {
-              id: '1',
-              name: 'Joe Smith'
-            }
-          ]
-        }
-      },
-      Person: {
-        __isTypeOf() {
-          return 'Person'
-        }
-      },
-      Animal: {
-        __isTypeOf() {
-          return 'Animal';
-        }
-      }
-    }
-  });
-
-  const parent = new GraphQLComponent({
-    imports: [
-      child
-    ]
-  });
-
-  const result = await graphql.execute({
-    document: gql`
-      query {
-        things {
-          id
-          ... on Person {
-            name
-          }
-        }
-      }
-    `,
-    schema: parent.schema,
-    contextValue: {}
-  });
-
-  t.deepEquals(result.data.things, [{ id: '1', name: 'Joe Smith', __typename: 'Person' }], 'interface type resolved')
-  t.notOk(result.errors, 'no errors');
-  t.end();
-});
