+++ lib/index.js
@@ -6,10 +6,11 @@ const { mergeResolvers, mergeTypeDefs } = require('@graphql-tools/merge');
 const { makeExecutableSchema } = require('@graphql-tools/schema');
 const { addMocksToSchema } = require('@graphql-tools/mock');
 const { SchemaDirectiveVisitor } = require('@graphql-tools/utils');
-const { wrapResolvers, delegateToComponent } = require('./resolvers');
+const { wrapResolvers } = require('./resolvers');
 const { buildDependencyTree } = require('./imports');
 const { wrapContext, createContext } = require('./context');
 const { createDataSourceInjection } = require('./datasource');
+const { delegateToComponent } = require('./delegate');
 const cuid = require('cuid');
 
 const debug = require('debug')('graphql-component:schema');
