+++ lib/resolvers/__tests__.js
@@ -414,47 +414,6 @@ Test('importResolvers()', (t) => {
     st.notOk(importedResolvers.SomeType, 'non-root type (SomeType) is not imported')
     st.end();
   });
-
-  t.test(`import component's resolvers that contain an abstract type`, (st) => {
-    const component = new GraphQLComponent({
-      types: `
-        type Query {
-          thing: Thing
-        }
-        interface Thing {
-          id: ID
-        }
-        type A implements Thing {
-          id: ID
-          fieldForA: String
-        }
-        type B implements Thing {
-          id: ID
-          fieldForB: String
-        }
-      `,
-      resolvers: {
-        Query: {
-          thing() {
-            return { id: '1', fieldForA: 'a', fieldForB: 'b'};
-          }
-        },
-        Thing: {
-          __resolveType(thing) {
-            if (thing.fieldForA) {
-              return 'A';
-            } else if (thing.fieldForB) {
-              return 'B';
-            }
-          }
-        }
-      }
-    });
-
-    const importedResolvers = importResolvers(component);
-    st.ok(importedResolvers.Query.thing.__isProxy, 'Query.thing is a proxy');
-    st.end();
-  });
 })
